requirements.txt
backports.csv==1.0.7
beautifulsoup4==4.6.0
boto==2.49.0
captcha==0.3
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.5
django-colorful==1.3
django-countries==5.5
django-crispy-forms==1.7.2
django-djconfig==0.10.0
django-el-pagination==3.2.4
django-extensions==2.2.1
django-geojson==2.12.0
django-import-export==1.2.0
django-leaflet==0.24.0
django-modelcluster==4.4
django-post-office==3.2.1
django-shapes==0.2.0
django-social-share==1.3.2
django-storages==1.7.2
django-taggit==0.24.0
django-treebeard==4.3
djangorestframework==3.10.3
draftjs-exporter==2.1.6
et-xmlfile==1.0.1
geojson==2.5.0
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
jdcal==1.4.1
jsonfield==2.0.2
mistune==0.8.4
numpy==1.16.5
oauthlib==3.1.0
odfpy==1.4.0
openpyxl==2.6.3
Pillow==6.1.0
psycopg2==2.8.3
puput==1.0.5
python3-openid==3.1.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
spirit==2.0.1
sqlparse==0.3.0
static3==0.7.0
tablib==0.13.0
Unidecode==1.1.1
urllib3==1.25.3
wagtail==2.6.1
webencodings==0.5.1
Willow==1.1
wincertstore==0.2
xlrd==1.2.0
xlwt==1.3.0
